fix: telegram special characters encoding
